requirements.txt
absl-py
blessings
cachetools
certifi
charset-normalizer
google-auth
google-auth-oauthlib
gpustat
GitPython
grpcio
h5py
idna
imageio
imutils
Markdown
matplotlib
numpy
nvidia-ml-py3
oauthlib
opencv-python
omegaconf
Pillow>=8.3.2
protobuf
psutil
pyasn1
pyasn1-modules
requests
requests-oauthlib
rich
rsa
six
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
torch==1.12.1+cu113 
torchvision==0.13.1+cu113
typing-extensions
urllib3
Werkzeug
coloredlogs
slackweb
scipy
aggdraw
sh
pysen
mypy
isort
flake8
black
pandas
seaborn
